Move Ms instrument handling code to separate files (MIDI import)
